--- tests/root.mutation.graphql
@@ -1,12 +1,0 @@
-type Mutation {
-    # Define a mutation register, which uses the Inputs defined in user.graphql, and uses the ResponsePayload as return
-    register (
-        input: RegistrationInput!
-    ): ResponsePayload!
-}
-
-# Custom resturn for response
-type ResponsePayload {
-    status: Boolean!
-    id: ID
-}
