--- tests/root.query.graphql
@@ -1,10 +1,0 @@
-type Query {
-  # Define the query that is used in graphql language, what inputs it can take,
-  # what BaseResult it should use.
-  # Directives can be added at the end of either a field or the BaseResult
-  # restricting the return of either field or the whole object. 
-  users(
-    paging: PagingInput
-    distinct: Boolean
-  ): User!
-}
