+++ tests/test_kafka.py
