+++ tests/test_routing.py
@@ -62,6 +62,3 @@ print(app.routes)
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=9000, workers=1)
-
-
-
