--- tests/tests-conftest.py
@@ -1,3 +1,0 @@
-import os
-
-os.environ["STARCLI_FOLDER"] = "tests"
